+++ src/main/java/com/financialapp/repository/RewardRepository.java
@@ -3,5 +3,5 @@ package com.financialapp.repository;
 import com.financialapp.model.Reward;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RewardRepository extends JpaRepository<Reward, Integer> {
+public interface RewardRepository extends JpaRepository<Reward, Long> {
 }
